task3: (fixes) tried to fix markup, moved pipes into 'pipes' folder, rewrited url paths for assets files
