+++ src/app/modules/core/core.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HeaderComponent } from './header/header.component';
 import { FooterComponent } from './footer/footer.component';
-import { DurationPipe } from 'app/modules/core/duration-pipe/duration.pipe';
+import { DurationPipe } from 'app/pipes/duration-pipe/duration.pipe';
 import { ConfirmModalComponent } from 'app/modules/core/confirm-modal/confirm-modal.component';
 import { AppRoutingModule } from 'app/app-routing.module';
