+++ src/app/modules/core/footer/footer.component.scss
@@ -6,7 +6,7 @@
   align-items: center;
   padding: 0 85px 0;
   background-color: $charcoal-grey;
-  font-family: 'Source Sans Pro';
+  font-family: Source Sans Pro;
   font-size: 14px;
   line-height: 1.71;
   color: $silver;
