+++ src/app/modules/core/header/header.component.scss
@@ -17,7 +17,7 @@
       width: 32px;
       height: 32px;
       margin: 0 10px 0 0;
-      background: url('app/../assets/logo.svg') no-repeat center;
+      background: url('assets/logo.svg') no-repeat center;
     }
 
     .logo-text {
@@ -35,7 +35,7 @@
 
     .login-logoff-text {
       margin-left: 10px;
-      font-family: 'Source Sans Pro';
+      font-family: Source Sans Pro;
       font-size: 14px;
       color: $white;
     }
@@ -48,7 +48,7 @@
       &__icon {
         width: 10px;
         height: 12px;
-        background: url('app/../assets/ic-user.svg') no-repeat center;
+        background: url('assets/ic-user.svg') no-repeat center;
       }
     }
 
@@ -61,7 +61,7 @@
       &__icon {
         width: 16px;
         height: 12px;
-        background: url('app/../assets/ic-exit.svg') no-repeat center;
+        background: url('assets/ic-exit.svg') no-repeat center;
       }
     }
   }
