+++ src/app/modules/courses-page/breadcrumbs/breadcrumbs.component.scss
@@ -6,7 +6,7 @@
   padding: 0 0 0 85px;
   height: 64px;
   background-color: $white;
-  font-family: 'Source Sans Pro';
+  font-family: Source Sans Pro;
   font-size: 14px;
   font-weight: 600;
   line-height: 1.71;
