+++ src/app/modules/courses-page/course-item/course-item.component.scss
@@ -15,7 +15,7 @@
 
     .course-title {
       margin-right: 5px;
-      font-family: 'Source Sans Pro';
+      font-family: Source Sans Pro;
       font-size: 24px;
       font-weight: 600;
       line-height: 1.25;
@@ -42,7 +42,7 @@
         -webkit-box-orient: vertical;
         -webkit-line-clamp: 3;
         overflow: hidden;
-        font-family: 'Source Sans Pro';
+        font-family: Source Sans Pro;
         font-size: 14px;
         line-height: 1.71;
       }
@@ -69,7 +69,7 @@
         &-text {
           height: 15px;
           margin-left: 8px;
-          font-family: 'Source Sans Pro';
+          font-family: Source Sans Pro;
           font-size: 12px;
         }
       }
@@ -85,7 +85,7 @@
         justify-content: center;
         border-radius: 3px;
         background-color: $dark-sky-blue;
-        font-family: 'Source Sans Pro';
+        font-family: Source Sans Pro;
         font-size: 12px;
         font-weight: 600;
         line-height: 1.83;
