+++ src/app/modules/courses-page/courses-list/courses-list.component.scss
@@ -10,10 +10,9 @@
     display: flex;
     align-items: center;
     align-self: flex-end;
-    padding: 0 24px 0 18px;
     background-color: $dark-sky-blue;
     border-radius: 3px;
-    font-family: 'Source Sans Pro';
+    font-family: Source Sans Pro;
     font-size: 14px;
     font-weight: 600;
     line-height: 2.57;
@@ -27,9 +26,9 @@
 
     .add-icon {
       position: relative;
-      width: 18px;
+      width: 9px;
       height: 18px;
-      margin-right: 12px;
+      margin: 0 12px 0 18px;
 
       &::before,
       &::after {
@@ -42,7 +41,6 @@
       }
 
       &::after {
-        right: 7px;
         transform: rotate(90deg);
       }
     }
@@ -62,7 +60,7 @@
     background-color: $white;
 
     .load-more-button {
-      font-family: 'Source Sans Pro';
+      font-family: Source Sans Pro;
       font-size: 12px;
       line-height: 1.5;
       letter-spacing: 0.19px;
@@ -87,7 +85,7 @@
     border-radius: 6px;
     box-shadow: 0 1px 24px 0 $box-shadow2;
     background-color: $white;
-    font-family: 'Source Sans Pro';
+    font-family: Source Sans Pro;
     font-size: 14px;
     line-height: 1.5;
     letter-spacing: 0.19px;
