+++ src/app/pipes/filter-by-course-name-pipe/filter-by-course-name.pipe.ts
